Separated writers, updated api, separated tests
