+++ hydromt_sfincs/components/geometries/thin_dams.py
@@ -10,7 +10,7 @@ from shapely.geometry import LineString
 from hydromt import hydromt_step
 from hydromt.model.components import ModelComponent
 
-from hydromt_sfincs import utils
+from hydromt_sfincs import readers, utils, writers
 
 if TYPE_CHECKING:
     from hydromt_sfincs.sfincs import SfincsModel
@@ -103,7 +103,7 @@ class SfincsThinDams(ModelComponent):
             raise FileNotFoundError(f"Thin dams file not found: {abs_file_path}")
 
         # Read thd file
-        struct = utils.read_geoms(abs_file_path)  # =utils.py function
+        struct = readers.read_geoms(abs_file_path)  # =utils.py function
         gdf = utils.linestring2gdf(struct, crs=self.model.crs)  # =utils.py function
 
         # Add to self._data
@@ -140,15 +140,14 @@ class SfincsThinDams(ModelComponent):
         struct = utils.gdf2linestring(self.data)
 
         # Write to thd file
-        utils.write_geoms(abs_file_path, struct, stype="thd", fmt=fmt)
+        writers.write_geoms(abs_file_path, struct, stype="thd", fmt=fmt)
 
         # write also as geojson:
         if self.model.write_gis:
-            utils.write_vector(
+            writers.write_vector(
                 self.data,
                 name="thd",
                 root=join(self.root.path, "gis"),
-                logger=logger,
             )
 
     def set(self, gdf: gpd.GeoDataFrame, merge: bool = True):
