+++ hydromt_sfincs/components/geometries/wave_makers.py
@@ -9,7 +9,7 @@ from shapely.geometry import LineString
 
 from hydromt.model.components import ModelComponent
 
-from hydromt_sfincs import utils
+from hydromt_sfincs import readers, utils, writers
 
 if TYPE_CHECKING:
     from hydromt_sfincs.sfincs import SfincsModel
@@ -101,7 +101,7 @@ class SfincsWaveMakers(ModelComponent):
             raise FileNotFoundError(f"Wave makers file not found: {abs_file_path}")
 
         # Read wvm file
-        struct = utils.read_geoms(abs_file_path)
+        struct = readers.read_geoms(abs_file_path)
         gdf = utils.linestring2gdf(struct, crs=self.model.crs)
 
         # Add to self._data
@@ -135,15 +135,14 @@ class SfincsWaveMakers(ModelComponent):
         struct = utils.gdf2linestring(self.data)
 
         # Write to wvm file
-        utils.write_geoms(abs_file_path, struct, stype="wvm", fmt=fmt)
+        writers.write_geoms(abs_file_path, struct, stype="wvm", fmt=fmt)
 
         # write also as geojson:
         if self.model.write_gis:
-            utils.write_vector(
+            writers.write_vector(
                 self.data,
                 name="wvm",
                 root=join(self.root.path, "gis"),
-                logger=logger,
             )
 
     def set(self, gdf: gpd.GeoDataFrame, merge: bool = True):
